requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
ollama>=0.3.0
mistralai>=1.0.0
PyYAML>=6.0
requests>=2.31.0
